app/requirements.txt
click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==0.14.1
tensorflow==2.0.0-beta1
h5py==2.10.0
keras==2.3.1 
Flask-SQLAlchemy==2.4.0
psycopg2-binary==2.8.3
vaderSentiment==3.2.1
tweepy==3.9
requests